put the CFL criteria before looping over the energies
The CFL criteria and the reduction in dt was before done only in the CN scheme.
This led to some errors regarding the enery degradation that was not "updating"
fast enough. To avoid that the user had to estimate what dt was needed and to
give it as input. The results would be saved over that dt as well, leading to
very heavy data files and folders.

Now, the CFL criteria and the reduction in dt is done before starting to loop
over the energies. The user just needs to enter the "t_sampling" array over
which the data are to be saved. The program then takes care of calculating what
dt is actually needed for numerical accuracy and runs with it. After the
transport of e- is calculated, the Ie is downsampled to match the "t_sampling"
before saving.