improve the input from file function to handle non-matching time arrays
The function Ie_top_from_file now handles non-matching time arrays for incoming
flux. It can repeat or drop elements of the incoming flux to match the time
resolution of the simulation. It can also "cut" the flux in time if the data is
given over a longer time than the simulation, or on the opposite fill the flux
with zeros if the data is given over a smaller time than the simulation.