better calculations of dt and of the CFL_factor
Before this commit, dt was calculated by repeatedly dividing dt_sampling by 2
until the CFL criteria would be satisfied. Sometimes, this led to an
unnecessarily small dt as the only possible values dt could take were
dt_sampling divided by powers of 2 (so 2, 4, 8, 16, 32, 64, ...)

With this commit, the CFL factor is not limited to be powers of 2 and can take
any integer values (like 12 or 30 ...). This leads to coarser dt, improving
performance of the code.