replace PyCall by PythonCall
PythonCall is the current recommended package for interoperability with
Python. It is in active development and the big advantage compared to
PyCall is that it creates an isolated Conda environment for each Julia
project while PyCall is using the same for all Julia projects (which
can lead to issues/bugs).