fix issue with analysing very long simulations
For long simulations where the number XX of the "IeFlickering-XX.mat" files
exceeded 99, the functions performing some analysis on the data were loading
the data in the wrong order.

This is because the file names were sorted lexicographically, i.e.
IeFlickering-100.mat would be before IeFlickering-11.mat. As a result the
outputs from these functions were completely mixed in time.

This is fixed by sorting the filenames correctly before loading them.